requirements.txt
attrs==21.2.0
beautifulsoup4==4.10.0
bracex==2.1.1
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
click-option-group==0.5.3
colorama==0.4.4
coverage==6.0
dataclasses==0.6
Deprecated==1.2.13
Flask==2.0.2
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-core==2.2.1
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.42.0
grpcio-status==1.42.0
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==3.2.0
libcst==0.3.23
MarkupSafe==2.0.1
mypy-extensions==0.4.3
mysql-connector==2.2.9
numpy==1.21.4
packaging==21.0
pandas==1.3.4
peewee==3.14.4
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyGithub==1.55
PyJWT==2.1.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
PyYAML==6.0
requests==2.26.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
semgrep==0.67.0
six==1.16.0
soupsieve==2.3.1
tqdm==4.62.3
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
wcmatch==8.2
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.6.0
